[chore #160780758] : Fix error in tests, refactor
