+++ api/v2/views.py
@@ -6,7 +6,7 @@ from flask_restful import Api
 
 # local imports
 from api.v2 import APP
-from api.v2.resources import user
+from api.v2.resources.user import UserRegistration
 
 
 API = Api(APP)
@@ -14,4 +14,4 @@ API = Api(APP)
 # Base url common to all endpoints
 BASE_URL = '/api/v2'
 
-API.add_resource(user.UserRegistration, '{}/auth/signup'.format(BASE_URL))
+API.add_resource(UserRegistration, '{}/auth/signup'.format(BASE_URL))
